chore: remove redundant echo statement in flake.nix
Removes an unnecessary "Installing package in editable mode..." message from the `venv` activation script in `flake.nix`. This message was redundant as there is no installation in the shell hook. It is expected users run pip install -e . on their own once as then it is added to the venv.